+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Frame processing routines are member functions
 * Fix on_headers called twice from basic_parser_v1
 * Constrain parser_v1 constructor
+* Improve first line serialization
 * Refine Parser concept
 
 API Changes: